Fix build. Fix versions in nuget package
