--- NiL.JS.Portable/Core/Backward.cs
@@ -1,193 +1,0 @@
-﻿using System;
-using System.Reflection;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Runtime.InteropServices;
-
-namespace System
-{
-    internal enum TypeCode
-    {
-        Empty = 0,          // Null reference
-        Object = 1,         // Instance that isn't a value
-        DBNull = 2,         // Database null value
-        Boolean = 3,        // Boolean
-        Char = 4,           // Unicode character
-        SByte = 5,          // Signed 8-bit integer
-        Byte = 6,           // Unsigned 8-bit integer
-        Int16 = 7,          // Signed 16-bit integer
-        UInt16 = 8,         // Unsigned 16-bit integer
-        Int32 = 9,          // Signed 32-bit integer
-        UInt32 = 10,        // Unsigned 32-bit integer
-        Int64 = 11,         // Signed 64-bit integer
-        UInt64 = 12,        // Unsigned 64-bit integer
-        Single = 13,        // IEEE 32-bit float
-        Double = 14,        // IEEE 64-bit double
-        Decimal = 15,       // Decimal
-        DateTime = 16,      // DateTime
-        String = 18,        // Unicode character string
-    }
-
-    namespace Reflection
-    {
-        internal enum MemberTypes
-        {
-            Constructor = 1,
-            Event = 2,
-            Field = 4,
-            Method = 8,
-            Property = 16,
-            TypeInfo = 32,
-            Custom = 64,
-            NestedType = 128,
-            All = 191,
-        }
-    }
-
-    /// <summary>
-    /// Portable version only!
-    /// </summary>
-    internal static class PortableBackward
-    {
-        internal static ReadOnlyCollection<T> AsReadOnly<T>(this IList<T> self)
-        {
-            return new ReadOnlyCollection<T>(self);
-        }
-
-        internal static ReadOnlyCollection<T> AsReadOnly<T>(this List<T> self)
-        {
-            return new ReadOnlyCollection<T>(self);
-        }
-
-        internal static bool IsAssignableFrom(this Type self, Type sourceType)
-        {
-            return self.GetTypeInfo().IsAssignableFrom(sourceType.GetTypeInfo());
-        }
-
-        internal static bool IsSubclassOf(this Type self, Type sourceType)
-        {
-            return self != sourceType && self.GetTypeInfo().IsAssignableFrom(sourceType.GetTypeInfo());
-        }
-
-        internal static Attribute[] GetCustomAttributes(this Type self, Type attributeType, bool inherit)
-        {
-            return self.GetTypeInfo().GetCustomAttributes(attributeType, inherit).ToArray();
-        }
-
-        internal static bool IsDefined(this Type self, Type attributeType, bool inherit)
-        {
-            return self.GetTypeInfo().IsDefined(attributeType, inherit);
-        }
-        
-        internal static MemberTypes GetMemberType(this MemberInfo self)
-        {
-            if (self is ConstructorInfo)
-                return MemberTypes.Constructor;
-            if (self is EventInfo)
-                return MemberTypes.Event;
-            if (self is FieldInfo)
-                return MemberTypes.Field;
-            if (self is MethodInfo)
-                return MemberTypes.Method;
-            if (self is TypeInfo)
-                return MemberTypes.TypeInfo;
-            if (self is PropertyInfo)
-                return MemberTypes.Property;
-            return MemberTypes.Custom; // чёт своё, пускай сами разбираются
-        }
-
-        internal static MethodInfo GetGetMethod(this PropertyInfo self, bool fictive)
-        {
-            return self.GetMethod;
-        }
-
-        internal static MethodInfo GetSetMethod(this PropertyInfo self, bool fictive)
-        {
-            return self.SetMethod;
-        }
-
-        internal static MethodInfo GetAddMethod(this EventInfo self, bool fictive)
-        {
-            return self.AddMethod;
-        }
-
-        internal static MethodInfo GetGetMethod(this PropertyInfo self)
-        {
-            return self.GetMethod;
-        }
-
-        internal static MethodInfo GetSetMethod(this PropertyInfo self)
-        {
-            return self.SetMethod;
-        }
-
-        internal static MethodInfo GetAddMethod(this EventInfo self)
-        {
-            return self.AddMethod;
-        }
-
-        private static readonly Type[] _Types = 
-            {
-                null,
-                typeof(object),
-                Type.GetType("System.DBNull"),
-                typeof(bool),
-                typeof(char),
-                typeof(sbyte),
-                typeof(byte),
-                typeof(short),
-                typeof(ushort),
-                typeof(int),
-                typeof(uint),
-                typeof(long),
-                typeof(ulong),
-                typeof(float),
-                typeof(double),
-                typeof(decimal),
-                typeof(DateTime),
-                null,
-                typeof(string)
-            };
-
-        internal static TypeCode GetTypeCode(this Type type)
-        {
-            if (type == null)
-                return TypeCode.Empty;
-
-            if (type.GetTypeInfo().IsClass)
-            {
-                if (type == _Types[2])
-                    return TypeCode.DBNull;
-
-                if (type == typeof(string))
-                    return TypeCode.String;
-
-                return TypeCode.Object;
-            }
-
-            for (var i = 3; i < _Types.Length; i++)
-            {
-                if (_Types[i] == type)
-                    return (TypeCode)i;
-            }
-
-            return TypeCode.Object;
-        }
-
-        internal static Type GetInterface(this Type type, string name)
-        {
-            return type.GetTypeInfo().ImplementedInterfaces.First(x => x.Name == name);
-        }
-
-        internal static Type[] GetGenericArguments(this Type type)
-        {
-            return type.GenericTypeArguments;
-        }
-
-        internal static MethodInfo GetMethod(this Type type, string name, Type[] parameters)
-        {
-            return type.GetRuntimeMethod(name, parameters);
-        }
-    }
-}
