+++ NiL.JS/BaseLibrary/JSON.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Reflection;
 using System.Text;
+using NiL.JS.Backward;
 using NiL.JS.Core;
 using NiL.JS.Core.Interop;
