+++ NiL.JS/Core/Functions/MethodProxy.cs
@@ -4,12 +4,9 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Reflection.Emit;
+using NiL.JS.Backward;
 using NiL.JS.BaseLibrary;
 using NiL.JS.Core.Interop;
-
-#if NET40
-using NiL.JS.Backward;
-#endif
 
 namespace NiL.JS.Core.Functions
 {
