+++ NiL.JS/Core/Interop/JsonSerializersRegistry.cs
@@ -1,3 +1,4 @@
+using NiL.JS.Backward;
 using System;
 using System.Collections.Generic;
 using System.Reflection;
