+++ NiL.JS/NiL.JS.nuspec
@@ -17,7 +17,10 @@
   <files>
     <file src="..\NiL.JS.Portable\bin\Release\NiL.JS.dll" target="lib\portable-net45+win+wpa81+wp80" />
     <file src="..\NiL.JS.Portable\bin\Release\NiL.JS.pdb" target="lib\portable-net45+win+wpa81+wp80" />
-    <file src="\bin\release\nil.js.pdb" target="lib\net45" />
+    <file src="..\net461\nil.js.dll" target="lib\net461" />
+    <file src="..\net461\nil.js.pdb" target="lib\net461" />
+    <file src="..\net45\nil.js.dll" target="lib\net45" />
+    <file src="..\net45\nil.js.pdb" target="lib\net45" />
     <file src="..\net40\nil.js.dll" target="lib\net40" />
     <file src="..\net40\nil.js.pdb" target="lib\net40" />
   </files>
