+++ NiL.JS/Statements/ExportStatement.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using NiL.JS.Backward;
 using NiL.JS.Core;
 using NiL.JS.Expressions;
