+++ NiL.JS/Statements/ImportStatement.cs
@@ -1,4 +1,5 @@
-﻿using NiL.JS.Core;
+﻿using NiL.JS.Backward;
+using NiL.JS.Core;
 using System;
 using System.Collections.Generic;
 using System.Text;
